Remove GitHub Workflow for mirroring to Gitlab
There was an advent recently, the Pipelines in Gitlab took a very long
time to execute, and used up my entire quota of free runners. At first
I thought of just adding timeouts, but then I thought there might be
better approaches to integrating with Gitlab.

So I'm removing the Gitlab integration for now, maybe in the future
when quotas are restored I'll set up a more convenient and thoughtful
integration again.